We have a fully functioning script!
Everything is working as intended and it feels awesome. We are able to create a master password, generate new passwords and store them, and view them. Everything is encrypted so unless the user knows the password they can't access the info. Going to clean everything up do some last finishing touches and we should be good to go!